Extract useDeleteVideo hook and move playlist handlers into useWatchedTime
